first prototypal driving system
My UGV reacts to an incoming obstacle by moving away from it. This becomes a steering action away from obstacle. Plus it has knowledge of its right and left wing. Ugly code though, but it's working as show in the attached video (the camera is working in best possible environment).